requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
chromadb==0.4.15
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
streamlit==1.28.0
requests==2.31.0
pydantic==2.5.0
httpx==0.25.0
reportlab==4.0.4
plotly==5.17.0
pandas==2.1.3